% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_annotation.R
\name{mf_annotation}
\alias{mf_annotation}
\title{Deprecated - Plot an annotation}
\usage{
mf_annotation(
  x,
  txt,
  pos = "topright",
  cex = 0.8,
  col_arrow,
  col_txt,
  halo = FALSE,
  bg,
  s = 1,
  ...
)
}
\arguments{
\item{x}{an sf object with 1 row, a couple of coordinates (c(x, y)) or
"interactive"}

\item{txt}{the text to display}

\item{pos}{position of the text, one of "topleft", "topright", "bottomright",
"bottomleft" or "center"}

\item{cex}{size of the text}

\item{col_arrow}{arrow color}

\item{col_txt}{text color}

\item{halo}{add a halo around the text}

\item{bg}{halo color}

\item{s}{arrow size (min=1)}

\item{...}{further \link{text} arguments.}
}
\value{
No return value, an annotation is displayed.
}
\description{
This function is deprecated. Please use \link{mf_text} instead.
instead.

Plot an annotation on a map.
}
\note{
Annotations cannot be displayed on unprojected (long/lat) maps.
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mf_text(
  x = mtq[2, ],
  txt = "pos = 'bottomleft'\nline = 2\nclockwise = FALSE",
  pos = "bottomleft",
  offset = 6,
  clockwise = FALSE,
  line = 2,
  box = FALSE
)
mf_text(
  x = mtq[28, ],
  txt = "pos = 'topright'\nline = 3\nbox= TRUE\nclockwise = FALSE",
  pos = "topright",
  offset = 10,
  clockwise = FALSE,
  line = 3,
  halo = TRUE,
  align = "left"
)
}
