% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_arrow.R
\name{mf_arrow}
\alias{mf_arrow}
\title{Plot a north arrow}
\usage{
mf_arrow(
  pos = "topleft",
  col = getOption("mapsf.fg"),
  cex = 1,
  adj = c(0, 0),
  align,
  adjust
)
}
\arguments{
\item{pos}{position. It can be one of 'topleft', 'top','topright', 'right', 'bottomright', 'bottom','bottomleft', 'left', 'interactive' or a vector of two coordinates in map units (c(x, y))}

\item{col}{arrow color}

\item{cex}{arrow size}

\item{adj}{adjust the postion of the north arrow in x and y directions}

\item{align}{object of class \code{sf} or \code{sfc} used to adjust the
arrow to the real north}

\item{adjust}{deprecated, see align}
}
\value{
No return value, a north arrow is displayed.
}
\description{
Plot a north arrow.
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mf_arrow(pos = "topright")
}
