% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_credits.R
\name{mf_credits}
\alias{mf_credits}
\title{Plot credits}
\usage{
mf_credits(
  txt = "Source(s) & Author(s)",
  pos = "bottomleft",
  col,
  cex = 0.6,
  font = 3,
  bg = NA
)
}
\arguments{
\item{txt}{text of the credits, use '\\n' to add line breaks}

\item{pos}{position, one of 'bottomleft', 'bottomright' or 'rightbottom'}

\item{col}{color}

\item{cex}{cex of the credits}

\item{font}{font of the credits}

\item{bg}{background color}
}
\value{
No return value, credits are displayed.
}
\description{
Plot credits (sources, author, year...).
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mf_credits(txt = "Author\nSources - Year")
}
