% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_frame.R
\name{mf_frame}
\alias{mf_frame}
\title{Plot a frame}
\usage{
mf_frame(extent = "map", col, lwd = 1.5, lty = 1, ...)
}
\arguments{
\item{extent}{type of frame, either 'map' or 'figure'}

\item{col}{line color}

\item{lwd}{line width}

\item{lty}{line type}

\item{...}{other arguments from \code{\link[graphics:box]{box}}}
}
\value{
No return value, a frame is displayed.
}
\description{
Plot a frame around an existing map.
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mf_title()
mf_frame(extent = "map")
mf_map(mtq)
mf_title()
mf_frame(extent = "figure")
}
