% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_get_leg_pos.R
\name{mf_get_leg_pos}
\alias{mf_get_leg_pos}
\title{Get the optimal position of a legend}
\usage{
mf_get_leg_pos(x, n = 1)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{n}{number of positions to get (1 or 2)}
}
\value{
A vector of position is returned
}
\description{
Find the optimal for one or two legends. The optimal position is
a position that minimizes overlap between a spatial object and a legend.
}
\examples{
mtq <- mf_get_mtq()
mf_get_leg_pos(mtq)
}
\keyword{internal}
