% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_theme.R
\name{mf_get_theme_value}
\alias{mf_get_theme_value}
\title{Get parameters of the current theme}
\usage{
mf_get_theme_value(arg)
}
\arguments{
\item{arg}{name of an argument of \link{mf_theme}}
}
\value{
\code{mf_get_theme_value} returns the value of a specific argument of
the current theme.
}
\description{
\code{mf_get_theme_value} is useful to access the value of a specific argument of
the current theme.
}
\examples{

# To obtain the value of one of the current theme argument:
mf_get_theme_value("foreground")
mf_get_theme_value("mar")
mf_get_theme_value("pal_seq")
}
\seealso{
\code{\link[=mf_theme]{mf_theme()}}
}
