% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_layout.R
\name{mf_layout}
\alias{mf_layout}
\title{Plot a map layout}
\usage{
mf_layout(
  title = "Map Title",
  credits = "Authors & Sources",
  scale = TRUE,
  arrow = TRUE,
  frame = FALSE
)
}
\arguments{
\item{title}{title of the map}

\item{credits}{credits}

\item{scale}{display a scale bar}

\item{arrow}{display an arrow}

\item{frame}{display a frame}
}
\description{
Plot a map layout (title, credits, scalebar,
north arrow, frame).
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mf_layout()
}
