% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_logo.R
\name{mf_logo}
\alias{mf_logo}
\title{Plot a logo on a map}
\usage{
mf_logo(filename, pos = "bottomright", cex = 1, adj = c(0, 0))
}
\arguments{
\item{filename}{filename of the logo image, PNG or JPG/JPEG format.}

\item{pos}{position of the logo, one of 'topleft', 'top','topright',
'right', 'bottomright', 'bottom', 'bottomleft', 'left' or a vector of two
coordinates in map units (c(x, y)). Use 'interactive' to choose the
legend position by clicking on the map.}

\item{cex}{amount by which the logo width should be magnified
or reduced relative to the default}

\item{adj}{adjust the position of the logo in x and y directions}
}
\value{
No return value, a background image is displayed.
}
\description{
The logo can be a PNG or JPG/JPEG file.
}
\examples{
m <- mf_get_mtq()
mf_map(m)
mf_scale()
logo <- system.file("img", "Rlogo.png", package = "png")
mf_logo(logo, pos = "bottomleft", adj = c(0, 4))
mf_credits()
}
