% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_map_help.R
\name{mf_map-prop}
\alias{mf_map-prop}
\title{Plot proportional symbols}
\arguments{
\item{x}{object of class \code{sf}}

\item{var}{name of the variable to map}

\item{type}{"prop"}

\item{inches}{size of the largest symbol in inches (radius for circles,
half width for squares)}

\item{val_max}{maximum value corresponding to the largest symbol or line}

\item{lwd_max}{width of the largest line}

\item{symbol}{type of proportional symbols, either "circle" or "square"}

\item{col}{color of the proportional symbols or lines, a hex code or
a color name given by \link{colors}.
The default color is the highlight color (see \link{mf_get_theme_value}).}

\item{border}{border color for proportional symbols, a hex code or
color name given by \link{colors}. The default color the background color (see
\link{mf_get_theme_value}).}

\item{lwd}{border width of proportional symbols}

\item{alpha, expandBB, extent, bg, add}{arguments described in \link{mf_map}}

\item{leg_*}{legend arguments described in \link{mf_map}}
}
\value{
x is (invisibly) returned.
}
\description{
With the \strong{prop} map type, \code{mf_map()} displays symbols (squares or circles)
with areas proportional to a quantitative variable (stocks).

For polygons, centroids are used to plot proportional symbols.
\subsection{Usage}{

For polygons and points:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mf_map(x, var, type = "prop",
       inches = 0.3, val_max, symbol, col, border, lwd,
       expandBB, extent, bg, alpha, add = FALSE, leg_*)

}\if{html}{\out{</div>}}

For lines:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mf_map(x, var, type = "prop",
       val_max, lwd_max = 20, col,
       expandBB, extent, bg, alpha, add = FALSE, leg_*)
}\if{html}{\out{</div>}}
}
}
\examples{
mtq <- mf_get_mtq()
flows <- mf_get_mtq("lines")
mtq <- mf_get_mtq()
mf_map(mtq)
mf_map(mtq, "POP", "prop",
  inches = .4, leg_title = "Population",
  leg_pos = "topright"
)
mf_map(flows, "fij", "prop",
  lwd_max = 10, col = "steelblue2",
  leg_pos = "right"
)
}
\seealso{
\code{\link[=mf_map]{mf_map()}}
}
