% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_map_help.R
\name{mf_map-symb_choro}
\alias{mf_map-symb_choro}
\title{Plot symbols with choropleth coloration}
\arguments{
\item{x}{object of class \code{sf} (polygons or points)}

\item{var}{names of the variables to map. The first value refers to the
symbols categories, the second one to the choropleth coloration.}

\item{type}{"symb_choro"}

\item{pch}{a vector of types of symbols, see \link{pch}.
The length of \code{pch} should match the number of modalities.}

\item{cex}{a vector of sizes for symbols.
The length of \code{cex} should match the number of modalities.}

\item{lwd}{border width of symbols}

\item{border}{border color for symbols, a hex code or color name given
by \link{colors}. The default color is the background color (see
\link{mf_get_theme_value}).}

\item{val_order}{modalities order in the legend, a character vector that
matches \code{var} modalities. Default to alphabetic order of modalities.}

\item{pal}{a set of colors (hex codes) or a palette name. Palette names can
be obtained with \link{hcl.pals}. The default palette is the pal_seq palette
(see \link{mf_get_theme_value}).}

\item{rev}{if \code{pal} is a palette name, whether the ordering of the colors
should be reversed (TRUE) or not (FALSE)}

\item{breaks}{either a numeric vector with the actual breaks, or a
classification method name. The main methods are 'quantile',
'equal', 'msd', 'ckmeans' (natural breaks), 'Q6' and 'geom'.
See \link{mf_get_breaks} for details.}

\item{nbreaks}{number of classes}

\item{pch_na}{type of symbol for missing values, see \link{pch}}

\item{cex_na}{size of symbol for missing values}

\item{col_na}{color for missing values, a hex code or
a color name given by \link{colors}}

\item{alpha, expandBB, extent, bg, add}{arguments described in \link{mf_map}}

\item{leg_*}{legend arguments described in \link{mf_map}. See details for
arguments with two values.}
}
\value{
x is (invisibly) returned.
}
\description{
\code{mf_map()} with \strong{symb_choro} type creates symbols that reflect modalities
of a first qualitative variable and colored to reflect the classification of
a second variable.

This map types uses two variables and some arguments need to be set for both
variables (see Details).

For polygons, centroids are used to plot symbols. This map type
is not available for lines.
\subsection{Usage}{

For polygons and points:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mf_map(x, var, type = "symb_choro",
       pch, cex = 2, lwd = 0.7, border, val_order,
       pal, rev = FALSE, breaks = "quantile", nbreaks,
       pch_na = 4, cex_na = 1, col_na = "white",
       alpha, expandBB, extent, bg, add = TRUE, leg_*)
}\if{html}{\out{</div>}}
}
}
\details{
Legend arguments that need two values are: 'leg_title',
'leg_no_data'. The first value refers to the symbols legend, the second
one to the choropleth legend.
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mtq$STATUS[4] <- NA
mf_map(mtq, c("STATUS", "MED"),
  type = "symb_choro", lwd = 1,
  pal = "Reds 3", breaks = "quantile", nbreaks = 4,
  cex = c(2, 1, 1), pch = c(20, 21, 23), pch_na = 22,
  leg_pos = "topright", border = "white",
  val_order = c("Prefecture", "Sub-prefecture", "Simple municipality")
)
}
\seealso{
\code{\link[=mf_map]{mf_map()}}, \link{mf_map-symb}, \link{mf_map-choro}, \code{\link[=mf_distr]{mf_distr()}},
\code{\link[=mf_get_breaks]{mf_get_breaks()}}, \code{\link[=mf_get_pal]{mf_get_pal()}}
}
