% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_map_help.R
\name{mf_map_base}
\alias{mf_map_base}
\title{Plot a base map}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{col}{a color, hex code or color name given by \link{colors}.
The default color for polygons is the foreground color, the default color for
points and lines is the highlight color (see \link{mf_theme}).}

\item{border}{border color for polygons and points symbols, hex code or
color name given by \link{colors}. The default color for polygon is the
highlight color, the default color for points is the foreground color (see
\link{mf_theme}).}

\item{lwd}{border width for polygons and points symbols, lines width}

\item{lty}{type of line for polygons borders and lines}

\item{pch}{type of symbol to use for points, see \link{pch}}

\item{cex}{symbols size, 2 means 2 times bigger}

\item{alpha, expandBB, extent, bg, add}{arguments described in \link{mf_map}}
}
\value{
x is (invisibly) returned.
}
\description{
\code{mf_map()} can be used to display geographic layers (\code{sf} objects), using
the default map type \strong{base}.
\subsection{Usage}{

For polygons:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mf_map(x, col, border, lwd = .7, lty = 1,
       alpha, expandBB, extent, bg, add = FALSE)
}\if{html}{\out{</div>}}

For points:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mf_map(x, col, border, pch = 20, cex = 1, lwd = .7,
       alpha, expandBB, extent, bg, add = FALSE)
}\if{html}{\out{</div>}}

For lines:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mf_map(x, col, lwd = .7, lty = 1,
       alpha, expandBB, extent, bg, add = FALSE)
}\if{html}{\out{</div>}}
}
}
\examples{
mtq <- mf_get_mtq()
pts <- mf_get_mtq("points")
flows <- mf_get_mtq("lines")
mf_map(mtq, lty = 3)
mf_map(pts, col = "red", border = "white", pch = 21, add = TRUE)
mf_map(flows, col = "coral", lwd = 2, add = TRUE)
}
\seealso{
\code{\link[=mf_map]{mf_map()}}
}
