% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_map_help.R
\name{mf_map_symb}
\alias{mf_map_symb}
\title{Plot symbols}
\arguments{
\item{x}{object of class \code{sf} (polygons or points)}

\item{var}{name of the variable to map}

\item{type}{"symb"}

\item{pch}{a vector of types of symbols, see \link{pch}.
The length of \code{pch} should match the number of modalities.}

\item{cex}{a vector of sizes for symbols.
The length of \code{cex} should match the number of modalities.}

\item{lwd}{border width of symbols}

\item{pal}{a set of colors (hex codes) or a palette name. Palette names can
be obtained with \link{hcl.pals}. The default palette is the pal_quali palette
(see \link{mf_get_theme_value}).}

\item{rev}{if \code{pal} is a palette name, whether the ordering of the colors
should be reversed (TRUE) or not (FALSE)}

\item{border}{border color for symbols, a hex code or color name given
by \link{colors}. The default color is the background color (see
\link{mf_get_theme_value}).}

\item{val_order}{modalities order in the legend, a character vector that
matches \code{var} modalities. Default to alphabetic order of modalities.}

\item{pch_na}{type of symbol for missing values, see \link{pch}}

\item{cex_na}{size of symbol for missing values}

\item{col_na}{color for missing values, a hex code or
a color name given by \link{colors}}
}
\value{
x is (invisibly) returned.
}
\description{
\code{mf_map()} can use symbols to display qualitative data, using \strong{symb}
map type.

For polygons, centroids are used to plot graduated symbols. This map type
is not available for lines.
\subsection{Usage}{

For polygons and points:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mf_map(x, var, type = "symb",
       pch, cex = 2, lwd = 0.7, pal, rev = FALSE, border,
       val_order,
       col_na = "grey", pch_na = 4, cex_na = 1,
       alpha, expandBB, extent, bg, add = TRUE, leg_*)
}\if{html}{\out{</div>}}
}
}
\examples{
mtq <- mf_get_mtq()
mtq$STATUS[3] <- NA
mf_map(mtq)
mf_map(mtq, "STATUS", "symb",
  pal = "Berlin", border = "white", lwd = 1,
  cex = c(4, 3, 2), pch = c(21:23), col_na = "red",
  val_order = c("Prefecture", "Sub-prefecture", "Simple municipality"),
  leg_title = ""
)
}
