% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_svg.R
\name{mf_svg}
\alias{mf_svg}
\title{Export a map in svg format}
\usage{
mf_svg(
  x,
  filename = "map.svg",
  width,
  height,
  expandBB = rep(0, 4),
  svglite = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{SpatRaster}}

\item{filename}{path to the exported file.}

\item{width}{width of the figure (inches)}

\item{height}{height of the figure (inches)}

\item{expandBB}{fractional values to expand the bounding box with, in each
direction (bottom, left, top, right)}

\item{svglite}{if TRUE, the export is done with the \code{svglite} package
if installed}

\item{...}{further parameters}
}
\value{
No return value, an SVG device is initiated.
}
\description{
Export a map with the extent of a spatial object.\cr
The map is exported in SVG format.\cr
If only one of \code{width} or \code{height} is set, \code{mf_export} uses
the width/height ratio of \code{x} bounding box to find a matching ratio for
the export.\cr
Use \code{dev.off} to finish the export (see Examples).
}
\examples{
mtq <- mf_get_mtq()
(filename <- tempfile(fileext = ".svg"))
mf_svg(mtq, filename = filename)
mf_map(mtq)
mf_title()
dev.off()
}
