% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_title.R
\name{mf_title}
\alias{mf_title}
\title{Plot a title}
\usage{
mf_title(txt = "Map Title", pos, tab, bg, fg, cex, line, font, inner, banner)
}
\arguments{
\item{txt}{title text}

\item{pos}{position, one of 'left', 'center', 'right'}

\item{tab}{if TRUE the title is displayed as a tab}

\item{bg}{background color of the title, hex code or color name given by
\link{colors}}

\item{fg}{foreground color of the title, hex code or color name given by
\link{colors}}

\item{cex}{size of the title}

\item{line}{number of margin lines used for the title}

\item{font}{font of the title}

\item{inner}{if TRUE the title is displayed inside the plot area;
if FALSE the title is displayed in the top margin}

\item{banner}{if TRUE the title is displayed as a banner}
}
\value{
No return value, a title is displayed.
}
\description{
Default values for \code{mf_title()} arguments are set with the map theme
(see \link{mf_theme}).
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mf_title()
}
