% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mapsf-deprecated}
\alias{mapsf-deprecated}
\title{Deprecated functions in mapsf}
\description{
These functions and features still work but will be removed in the next
major version of the package.
\subsection{\code{mf_map} sub-functions}{

Instead of using the following deprecated functions, one can use \link{mf_map}
with the corresponding type:
\itemize{
\item \code{mf_base()} => \strong{base}
\item \code{mf_choro()} => \strong{choro}
\item \code{mf_prop()} => \strong{prop}
\item \code{mf_typo()} => \strong{typo}
\item \code{mf_symb()} => \strong{symb}
\item \code{mf_grad()} => \strong{grad}
\item \code{mf_prop_typo()} => \strong{prop_typo}
\item \code{mf_prop_choro()} => \strong{prop_choro}
\item \code{mf_symb_choro()} => \strong{symb_choro}
}
}

\subsection{\code{mf_init}}{

\code{mf_init} is deprecated. It is possible to use \code{mf_map} instead
(\code{mf_map(x, type = "base", col = NA, border = NA)}.

It is also possible to use the \code{extent} argument of \link{mf_map}.
}

\subsection{\code{mf_export}}{

\code{mf_export} is deprecated, use \link{mf_png} or \link{mf_svg} instead.
}

\subsection{\code{mf_annotation}}{

\code{mf_annotation} is deprecated, use \link{mf_text} instead.
}

\subsection{Double legends}{

The use of separated legends for map types \strong{prop_choro}, \strong{prop_typo} and
\strong{symb_choro} is deprecated. Use \code{leg_pos = NA} and \link{mf_legend} if
separated legends are needed.
}

\subsection{Theming system}{

In \link{mf_theme}, the following themes are deprecated: "default", "brutal",
"ink", "dark", "agolalight", "candy", "darkula", "iceberg", "green",
"nevermind", "jsk" and "barcelona".\cr
The following arguments are also deprecated: "bg", "fg", "tab",
"pos", "inner", "line", "cex" and "font".\cr

Although the map theming system has been radically changed in version 1.0.0
of the package, you can still use the old themes by referencing them by name.
If you need to use the \emph{pre} v1.0.0 default theme, set \code{x} to "default".\cr
If an old theme is set, only deprecated arguments are used and others are
ignored.\cr
If current and deprecated arguments are mixed, only deprecated arguments are
used and others are ignored.\cr
All references and usages of the old theming system will be removed in the
next major version.
}
}
