% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mapsf-deprecated}
\alias{mapsf-deprecated}
\title{Deprecated functions in mapsf}
\description{
These functions still work but will be removed (defunct) in the next
major version of the package.
}
\details{
Instead of using the following functions, one can use \code{mf_map()} with the
corresponding type:
\itemize{
\item \verb{mf_base() => mf_map(x, type  = "base")}
\item \verb{mf_choro() => mf_map(x, var, type = "choro")}
\item \verb{mf_prop() => mf_map(x, var, type = "prop")}
\item \verb{mf_typo() => mf_map(x, var, type = "typo")}
\item \verb{mf_symb() => mf_map(x, var, type = "symb")}
\item \verb{mf_grad() => mf_map(x, var, type = "grad")}
\item \verb{mf_prop_typo() => mf_map(x, var, type = "prop_typo")}
\item \verb{mf_prop_choro() => mf_map(x, var, type = "prop_choro")}
\item \verb{mf_symb_choro() => mf_map(x, var, type = "symb_choro")}
\item \verb{mf_init() => mf_map(x, type = "base", col = NA, border = NA)}
}
}
