% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mapsf}
\alias{mapsf}
\title{Package description}
\description{
Create maps with simple features. mapsf helps to map sf objects and offers
features that improve the graphic presentation of maps (scale bar,
north arrow, title or legend).
}
