% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_base.R
\name{mf_base}
\alias{mf_base}
\title{Plot an sf object}
\usage{
mf_base(x, col = "grey80", border = "grey20", lwd = 0.7, pch = 20, add = FALSE)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{col}{color}

\item{border}{border color}

\item{lwd}{border width}

\item{pch}{pch (point type) for symbols}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}
}
\value{
x is (invisibly) returned.
}
\description{
Plot an sf object. This is mostly a wrapper around
\code{plot(st_geometry(x), ...)}.
}
\examples{
library(sf)
mtq <- mf_get_mtq()
mf_base(mtq)
mf_base(mtq, col = "blue")
}
\keyword{internal}
