% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_get_mtq.R
\name{mf_get_mtq}
\alias{mf_get_mtq}
\title{Get the 'mtq' dataset}
\usage{
mf_get_mtq()
}
\value{
an sf object of Martinique municipalities
}
\description{
Import the mtq dataset (Martinique municipalities).
}
\details{
This a wrapper around
\code{st_read(system.file("gpkg/mtq.gpkg", package = "mapsf"),quiet = TRUE)}.
}
\examples{
mtq <- mf_get_mtq()
}
