% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_raster.R
\name{mf_raster}
\alias{mf_raster}
\title{Plot a raster}
\usage{
mf_raster(x, add = FALSE, ...)
}
\arguments{
\item{x}{a SpatRaster}

\item{add}{whether to add the layer to an existing plot (TRUE) or
not (FALSE).}

\item{...}{bgalpha, interpolate, or other arguments passed to be passed to
\code{\link[terra:plotRGB]{plotRGB}} or  \code{\link[terra:plot]{plot}}}
}
\value{
No return value, a map is displayed.
}
\description{
Plot a raster object (SpatRaster from terra).
}
\examples{
if (require("terra")) {
  r <- rast(system.file("ex/elev.tif", package = "terra"))
  mf_raster(r)
}
}
