% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_shadow.R
\name{mf_shadow}
\alias{mf_shadow}
\title{Plot a shadow}
\usage{
mf_shadow(
  x,
  col,
  cex = 1,
  add = FALSE,
  extent = x,
  bg,
  expandBB = rep(0.04, 4)
)
}
\arguments{
\item{x}{an sf or sfc polygon object}

\item{col}{shadow color, hex code or color name given by
\link{colors}.
The default color is the highlight color (see \link{mf_theme}).}

\item{cex}{shadow extent}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}

\item{extent}{\code{sf} object used to define the map extent; defaults to \code{x}.
\code{extent} and \code{x} must use the same CRS.}

\item{bg}{background color of the map, hex code or color name given by
\link{colors}, ignored if \code{add = TRUE}}

\item{expandBB}{expension of the map area in each direction (bottom, left,
top, right). The expension is expressed as a share of \code{x} width
(for left and right values) or a share of \code{x} height (for bottom and top
values).}
}
\value{
x is (invisibly) returned.
}
\description{
Plot the shadow of a polygon layer.
}
\examples{
mtq <- mf_get_mtq()
mf_shadow(mtq)
mf_map(mtq, add = TRUE)
}
