% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_shadow.R
\name{mf_shadow}
\alias{mf_shadow}
\title{Plot a shadow}
\usage{
mf_shadow(x, col = "grey50", cex = 1, add = FALSE)
}
\arguments{
\item{x}{an sf or sfc polygon object}

\item{col}{shadow color}

\item{cex}{shadow extent}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}
}
\value{
x is (invisibly) returned.
}
\description{
Plot the shadow of a polygon layer.
}
\examples{
mtq <- mf_get_mtq()
mf_shadow(mtq)
mf_map(mtq, add = TRUE)
}
