% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_theme.R
\name{mf_theme}
\alias{mf_theme}
\title{Set a theme}
\usage{
mf_theme(x)
}
\arguments{
\item{x}{name of a map theme. One of "default", "brutal", "ink",
"dark", "agolalight", "candy", "darkula", "iceberg", "green", "nevermind",
"jsk", "barcelona".
It is also possible to set a custom theme (see Examples).
The parameters set by this function are the figure margins, background and
foreground colors and some \link{mf_title} options.
Use \code{mf_theme('default')} to reset theme settings.}
}
\value{
The (invisible) list of parameters of the current theme.
}
\description{
This function set a color theme.
}
\examples{
mtq <- mf_get_mtq()
mf_theme("green")
mf_map(mtq)
mf_title()

custom <- list(
  name = "custom",
  bg = "green",
  fg = "red",
  mar = c(2, 2, 2, 2),
  tab = TRUE,
  pos = "center",
  inner = TRUE,
  line = 2,
  cex = 1.5,
  font = 3
)
mf_theme(custom)
mf_map(mtq)
mf_title()
mf_theme("default")
}
