% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_title.R
\name{mf_title}
\alias{mf_title}
\title{Plot a title}
\usage{
mf_title(txt = "Map Title", pos, tab, bg, fg, cex, line, font, inner)
}
\arguments{
\item{txt}{title text}

\item{pos}{position, one of 'left', 'center', 'right'}

\item{tab}{if TRUE the title is displayed as a 'tab'}

\item{bg}{background of the title}

\item{fg}{foreground of the title}

\item{cex}{cex of the title}

\item{line}{number of lines used for the title}

\item{font}{font of the title}

\item{inner}{if TRUE the title is displayed inside the plot area.}
}
\value{
No return value, a title is displayed.
}
\description{
Plot a title
}
\examples{
mtq <- mf_get_mtq()
mf_map(mtq)
mf_title()
}
