% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_worldmap.R
\name{mf_worldmap}
\alias{mf_worldmap}
\title{Plot a point on a world map}
\usage{
mf_worldmap(x, lon, lat, ...)
}
\arguments{
\item{x}{object of class \code{sf} or \code{sfc}}

\item{lon}{longitude}

\item{lat}{latitude}

\item{...}{further parameters to pass to points (cex, pch, col...).}
}
\value{
No return value, a world map is displayed.
}
\description{
Plot a point on a world map.
}
\note{
The main part of the code is stolen from @fzenoni
(\url{https://gist.github.com/fzenoni/ef23faf6d1ada5e4a91c9ef23b0ba2c1}).
}
\examples{
mtq <- mf_get_mtq()
mf_worldmap(mtq)
mf_worldmap(lon = 24, lat = 39)
}
